requirements.txt
numba==0.58.1
librosa==0.10.1
torchlibrosa==0.1.0
matplotlib
opencv-python==4.8.1.78
pillow==10.0.1
tqdm
pyyaml==5.4.1
numpy
scipy
scikit-learn
tensorboard
wandb
torcheval
yacs==0.1.8
moviepy==1.0.3
pandas
av